requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
black==24.3.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
commitizen==4.4.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-WTF==1.2.1
gunicorn
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pip==24.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==4.1.0
pycparser==2.21
pytest==8.0.2
python-dotenv==1.0.1
setuptools==70.0.0
requests==2.32.3
Werkzeug==3.0.6
wheel==0.41.3
WTForms==3.1.2
